+++ example2.py
@@ -10,8 +10,8 @@ scene: Scene = Scene()
 
 width = 640
 height = 480
-eye: Vec3f = jnp.array([2., 4., 1.])
-target: Vec3f = jnp.array([0., 0., 0.])
+eye: Vec3f = (2., 4., 1.)
+target: Vec3f = (0., 0., 0.)
 
 light: LightParameters = LightParameters()
 camera: CameraParameters = CameraParameters(
@@ -22,7 +22,7 @@ camera: CameraParameters = CameraParameters(
 )
 
 texture: Texture = build_texture_from_PyTinyrenderer(
-    jnp.array((
+    (
         255,
         255,
         255,  # White
@@ -35,7 +35,7 @@ texture: Texture = build_texture_from_PyTinyrenderer(
         0,
         0,
         255  # Blue
-    )),
+    ),
     2,
     2,
 ) / 255.0
@@ -62,23 +62,19 @@ uvs = jnp.array([
 ])
 
 indices = jnp.array([[0, 1, 2], [0, 2, 3]])
-model: Model = Model(
+model: Model = Model.create(
     verts=vertices,
     norms=normals,
     uvs=uvs,
     faces=indices,
-    faces_norm=indices,
-    faces_uv=indices,
     diffuse_map=texture,
-    # reference: https://github.com/erwincoumans/tinyrenderer/blob/89e8adafb35ecf5134e7b17b71b0f825939dc6d9/model.cpp#L215
-    specular_map=lax.full(texture.shape[:2], 2.0),
 )
 scene, plane_model = scene.add_model(model)
 
 scene, plane_instance_id = scene.add_object_instance(plane_model)
 scene = scene.set_object_orientation(
     plane_instance_id,
-    jnp.array([0, 0, 0, 1.]),
+    (0, 0, 0, 1.),
 )
 
 img = Renderer.get_camera_image(
@@ -95,6 +91,6 @@ rgb_array = lax.clamp(0., img * 255, 255.).astype(jnp.uint8)
 import matplotlib.pyplot as plt
 
 fig, ax = plt.subplots()
-ax.imshow(transpose_for_display(rgb_array), origin='lower')
+ax.imshow(transpose_for_display(rgb_array))
 
 plt.show()